Update GetAddressAmount() to always return 1000 for testing reasons

+++ src/Lykke.Service.Iota.Api/Services/IIotaService.cs
@@ -17,6 +17,6 @@ namespace Lykke.Service.Iota.Api.Services
 
         Task DeleteBroadcastAsync(IBroadcast broadcast);
 
-        Task<decimal> GetAddressBalance(string address);
+        Task<long> GetAddressBalance(string address);
     }
 }
